agentic_hr_project/backend/requirements.txt
Flask
Flask-CORS
langchain
langgraph
google-generativeai
faiss-cpu
pypdf
python-dotenv
pandas
numpy
sentence-transformers
requests